mods_panel: create mods_panel_list_inner.py and move class ModListItemInner to it
